+++ crates/radio-station/src/main.rs
@@ -27,6 +27,14 @@ pub async fn main() -> Result<(), anyhow::Error> {
                 .takes_value(true),
         )
         .arg(
+            clap::Arg::with_name("port")
+                .short("p")
+                .long("port")
+                .default_value("5002")
+                .help("Sets the SRS Port")
+                .takes_value(true),
+        )
+        .arg(
             clap::Arg::with_name("loop")
                 .short("l")
                 .long("loop")
@@ -43,6 +51,13 @@ pub async fn main() -> Result<(), anyhow::Error> {
     // Calling .unwrap() is safe here because "INPUT" is required
     let path = matches.value_of("PATH").unwrap();
     let should_loop = matches.is_present("loop");
+    let port = matches.value_of("port").unwrap();
+    let port = if let Ok(n) = u16::from_str(port) {
+        n
+    } else {
+        error!("The provided Port is not a valid number");
+        return Ok(());
+    };
     let freq = matches.value_of("frequency").unwrap();
     let freq = if let Ok(n) = u64::from_str(freq) {
         n
@@ -54,7 +69,7 @@ pub async fn main() -> Result<(), anyhow::Error> {
     let mut station = RadioStation::new("DCS Radio Station");
     station.set_frequency(freq);
     station.set_position(0.0, 0.0, 8000.);
-    station.set_port(5002);
+    station.set_port(port);
 
     info!("Start playing ...");
     station.play(path, should_loop).await?;